Updated budget and price updater
